% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noprobs.R
\name{any_problems}
\alias{any_problems}
\title{any_problems}
\usage{
any_problems(a_df)
}
\arguments{
\item{a_df}{}
}
\value{
a boolen indicating the presence of data in .warnings, .errors, or .messages
}
\description{
Check if there are any problems in unnested error data
}
\details{
.. content for \details{} ..
}
\author{
Miles McBain
}
