% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noprobs.R
\name{no_problem}
\alias{no_problem}
\title{no_problem}
\usage{
no_problem(expr, .result_col = ".result")
}
\arguments{
\item{expr}{an expression to evaluate}

\item{.result_col}{the name of the result column}
}
\value{
a one row tibble with error data
}
\description{
Capture errors warnings and messages in a tidy format
}
\details{
Outputs a 1 row tibble containing columns: .errors, .warnings, .messages and
the name supplied in .result_col

This is advantageous in tidy modeling workflows where you can use no_problem with map
}
\author{
Miles McBain
}
